@@ -2,10 +2,10 @@
 name = "codecrafters-redis"
 version = "0.1.0"
 authors = ["Codecrafters <hello@codecrafters.io>"]
-edition = "2021"
+edition = "2024"
 
 [dependencies]
-anyhow = "1.0.59"                                   # error handling
-bytes = "1.3.0"                                     # helps manage buffers
-thiserror = "1.0.32"                                # error handling
-tokio = { version = "1.23.0", features = ["full"] } # async networking
+anyhow = "1"                                   # error handling
+bytes = "1"                                    # helps manage buffers
+thiserror = "2"                                # error handling
+tokio = { version = "1", features = ["full"] } # async networking
